+++ Assets/Scripts/MVCBase/View.cs
@@ -4,11 +4,6 @@ namespace Assets.Scripts
 {
     public abstract class View : MonoBehaviour, IView
     {
-        protected IController _controller;
 
-        public void SetController(IController controller)
-        {
-            _controller = controller;
-        }
     }
 }
