+++ Assets/Scripts/PlayerCamera/PlayerCameraController.cs
@@ -4,27 +4,31 @@ namespace Assets.Scripts
 {
     public class PlayerCameraController : Controller<PlayerCameraModel, PlayerCameraView, PlayerCameraConfig>
     {
-        public PlayerCameraController(PlayerCameraModel model, PlayerCameraView view) : base(model, view)
+        private Transform _player;
+
+        public PlayerCameraController(PlayerCameraModel model, PlayerCameraView view, Transform player) : base(model, view)
         {
-            _view.OnInOffset += ChangePosition;
-            _model.OnPositionChanged += DisplayPosition;
+            _player = player;
+            GameStartup.OnFixedUpdate += ReadPlayerPosition;
+            _model.OnPositionChanged += _view.UpdatePosition;
         }
 
-        private void DisplayPosition(Vector3 position, int speed)
+        private void ReadPlayerPosition()
         {
-            _view.UpdatePosition(position, speed);
+            if (_view.transform.position != _player.transform.position)
+                ChangePosition(_player.position);
         }
 
         private void ChangePosition(Vector3 playerPosition)
         {
-            Vector3 newPosition = playerPosition + _view.Offset;
+            Vector3 newPosition = new Vector3(0, 0, playerPosition.z);
             _model.SetNewPosition(newPosition);
         }
 
         ~PlayerCameraController()
         {
-            _model.OnPositionChanged -= DisplayPosition;
-            _view.OnInOffset -= ChangePosition;
+            GameStartup.OnFixedUpdate -= ReadPlayerPosition;
+            _model.OnPositionChanged -= _view.UpdatePosition;
         }
     }
 }
